refactor(streamlit): extract prep spinner + Map tab to app_prep_map_ui
Refs #130. Implemented in #126.

Made-with: Cursor
